Crate xrb

X Rust Bindings

X Rust Bindings is a Rust library directly implementing the types and protocol messages of the X11 protocol specification. XRB is not a high-level API library, and it does not provide a direct connection to an X server, nor does it do anything else on its own. XRB’s development purpose is to provide a foundation for higher-level Rust API wrapper libraries. It is used by X.RS, the official accompanying API library for XRB.

To demonstrate the difference between X Rust Bindings and a higher-level API library, here is a comparison between the same protocol message in XRB and its higher-level equivalent in X.RS:

ConnectionInit request in XRB

/// A request to initiate a connection to the X server.
pub struct ConnectionInit<'a> {
    pub byte_order: ByteOrder,
    /// Should always be 11.
    pub protocol_major_version: u16,
    /// Should always be 0.
    pub protocol_minor_version: u16,
    pub auth_protocol_name: &'a str,
    pub auth_data: &'a str,
}

InitConnection request in X.RS

/// A request to initiate a connection to the X server.
pub struct InitConnection {}

Modules

bitmasks

error_handling 🔒

macros 🔒

notifications

proto 🔒

queries

replies

requests

serialization 🔒

Macros

bitmask

Creates a bitmask enum that can be serialized, deserialized, and compared.

doc

Generates a doc comment for the given tokens.

error

predefine

Defines a series of sequential constants, starting at 1.

res_id

Structs

Arc

Represents the arc of a circle.

Host

An X server host address with a Protocol family and the address itself.

Rect

A rectangle with x,y coordinates and width by height dimensions.

Enums

BitGravity

If the contents of a window should be kept when it is resized and where they should be placed.

Protocol

The protocol used to connection to a particular Host.

WinGravity

What to do with children of a window when that window is resized.

Constants

PROTOCOL_MAJOR_VERSION

The major version of the X protocol used in XRB. Should always be 11.

PROTOCOL_MINOR_VERSION

The minor version of the X protocol used in XRB. Should always be 0.

Type Definitions

Button

A button on the mouse, starting at 1.

Char2b

Represents a two-byte character in a UTF-16 string.

KeyCode

The encoding of a particular key actually present on the keyboard.

KeySym

The encoding of a particular key irrespective of its location or whether the keyboard includes it.

Point

Represents a point in 2D space; 2D coordinates. (x: i16, y: i16).

String16

A UTF-16 encoded string, represented as an array of two-byte characters.

Timestamp

A timestamp expressed in milliseconds, typically the time since the last server reset.