Module xrb::proto::atoms

Atom constants that are predefined in the X11 protocol.

The X11 protocol predefines the values of the following 68 Atoms for convenience. These Atoms always have the same value, across all libraries, all connections, and all clients. Other Atoms, i.e. Atoms not listed here, and which are not predefined, are dependent on the connection: the only way to know what these other Atoms’ values are for any given connection is to send InternAtom requests for them.

Once one InternAtom reply has been received, however, you can assume it will remain the same for the duration of the current connection.

Constants

ARC

Associated type: Arc

ATOM

Associated type: Atom

BITMAP

CAP_HEIGHT

CARDINAL

COLORMAP

Associated type: Colormap

COPYRIGHT

CURSOR

Associated type: Cursor

CUT_BUFFER0

CUT_BUFFER1

CUT_BUFFER2

CUT_BUFFER3

CUT_BUFFER4

CUT_BUFFER5

CUT_BUFFER6

CUT_BUFFER7

DRAWABLE

Associated type: Drawable

END_SPACE

FAMILY_NAME

FONT

Associated type: Font

FONT_NAME

FULL_NAME

INTEGER

ITALIC_ANGLE

MAX_SPACE

MIN_SPACE

NORM_SPACE

NOTICE

PIXMAP

Associated type: Pixmap

POINT

Associated type: Point

POINT_SIZE

PRIMARY

QUAD_WIDTH

RECTANGLE

Associated type: Rect

RESOLUTION

RESOURCE_MANAGER

RGB_BEST_MAP

RGB_BLUE_MAP

RGB_COLOR_MAP

RGB_DEFAULT_MAP

RGB_GRAY_MAP

RGB_GREEN_MAP

RGB_RED_MAP

SECONDARY

STRIKEOUT_ASCENT

STRIKEOUT_DESCENT

STRING

Associated type: String

SUBSCRIPT_X

SUBSCRIPT_Y

SUPERSCRIPT_X

SUPERSCRIPT_Y

UNDERLINE_POSITION

UNDERLINE_THICKNESS

VISUALID

Associated type: VisualId

WEIGHT

WINDOW

Associated type: Window

WM_CLASS

WM_CLIENT_MACHINE

WM_COMMAND

WM_HINTS

WM_ICON_NAME

WM_ICON_SIZE

WM_NAME

WM_NORMAL_HINTS

WM_SIZE_HINTS

WM_TRANSIENT_FOR

WM_ZOOM_HINTS

X_HEIGHT

Type Definitions

Atom

A unique ID corresponding to a String name registered with the X server.