Module xrb::proto::common

Structs

Arc

Represents the arc of a circle.

Host

An X server host address with a Protocol family and the address itself.

Rect

A rectangle with x,y coordinates and width by height dimensions.

Enums

BitGravity

If the contents of a window should be kept when it is resized and where they should be placed.

Protocol

The protocol used to connection to a particular Host.

WinGravity

What to do with children of a window when that window is resized.

Type Definitions

Button

A button on the mouse, starting at 1.

Char2b

Represents a two-byte character in a UTF-16 string.

KeyCode

The encoding of a particular key actually present on the keyboard.

KeySym

The encoding of a particular key irrespective of its location or whether the keyboard includes it.

Point

Represents a point in 2D space; 2D coordinates. (x: i16, y: i16).

String16

A UTF-16 encoded string, represented as an array of two-byte characters.

Timestamp

A timestamp expressed in milliseconds, typically the time since the last server reset.