Module xrb::proto::ids

Structs

Colormap

A resource ID referring to a colormap.

Cursor

A resource ID referring to an appearance of the cursor.

Font

A resource ID referring to a particular font. Implements Fontable.

GContext

A resource ID referring to a graphics context. Implements Fontable.

Pixmap

A resource ID referring to a pixmap. Implements Drawable.

VisualId

Window

A resource ID referring to a window. Implements Drawable.

Traits

Drawable

A trait object that represents either a Window or a Pixmap.

Fontable

A trait object that represents either a Font or a GContext.

ResId

A unique ID referring to a particular resource.