Trait xrb::proto::messages::requests::Request

pub trait Request {
    fn major_opcode() -> u8;
    fn length(&self) -> u16;

    fn minor_opcode() -> Option<u16> { ... }
}

Required Methods

fn major_opcode() -> u8

The major opcode is a unique identifier for this request.

Major opcodes 128 through 255 are reserved for extensions. An extension may contain multiple requests, and as such will represent multiple requests with the same major opcode. An extension may choose to encode an additional minor opcode in the metadata() byte.

fn length(&self) -> u16

The length of the request in units of four bytes.

Includes the length of the header, which is one unit of 4 bytes that contains the major_opcode(), metadata() byte, and these two length() bytes, as well as any additional data associated with the request.

The length() must be equal to the minimum length required to contain the request. That is, if the length of the request is not an exact multiple of 4 bytes, it should be rounded up to the nearest 4-byte unit by including however many padding bytes is necessary. Any unused padding bytes are not guaranteed to be zero; they may be set to anything.

Provided Methods

fn minor_opcode() -> Option<u16>

Extensions may specify minor opcodes to differentiate their requests.

It is up to the individual extension as to where the minor opcode is placed, though it is typically located in the ‘metadata’ byte of the request header, i.e. the second byte. This metadata byte may also be used for other purposes, however.