Trait xrb::serialization::Deserialize

pub trait Deserialize {
    fn deserialize(
        header: &[u8; 4], 
        buf: &mut impl Buf
    ) -> Result<Self, ReadError>
    where
        Self: Sized;

    fn deserialize_bytes(
        header: &[u8; 4], 
        bytes: &[u8]
    ) -> Result<Self, ReadError>
    where
        Self: Sized,
    { ... }
}

Deserializes a data structure from bytes. Values should implement ReadValue instead.

Implementations of Deserialize must be compatible with the existing implementation of Serialize for their type.

Required Methods

fn deserialize(header: &[u8; 4], buf: &mut impl Buf) -> Result<Self, ReadError> where
Self: Sized,

Deserializes the data structure from bytes in a Buf.

Note that bytes representing the type of data structure may not be present if they were used to determine which data structure’s deserialize function to call.

Should have zero side effects.

Provided Methods

fn deserialize_bytes(header: &[u8; 4], bytes: &[u8]) -> Result<Self, ReadError> where
Self: Sized,

Deserializes the data structure from bytes.

Do not implement this function: implement deserialize(buf) instead.