Trait xrb::serialization::write_value::WriteValue

pub trait WriteValue {
    fn write_1b(self) -> Result<u8, WriteError>;
    fn write_2b(self) -> Result<u16, WriteError>;
    fn write_4b(self) -> Result<u32, WriteError>;

    fn write_1b_to(self, buf: &mut impl BufMut) -> Result<(), WriteError>
    where
        Self: Sized,
    { ... }
    fn write_2b_to(self, buf: &mut impl BufMut) -> Result<(), WriteError>
    where
        Self: Sized,
    { ... }
    fn write_4b_to(self, buf: &mut impl BufMut) -> Result<(), WriteError>
    where
        Self: Sized,
    { ... }
}

Write a value as a 1-byte, 2-byte, or 4-byte unsigned integer (u8, u16, or u32).

A value is considered to be data that refers to one individual ‘piece’ of information. A value does not contain any other values.

A value is no more than four bytes in length, and can be expressed as a u8, u16, or u32 value.

The term value is used here to refer to the concept of a value or field: an Option of a a value can still be considered a value, because it is effectively just the value itself, even though it might technically have an unnamed field. The point is that WriteValue should not be implemented for a data structure, such as an object, a list, a message, though it might be implemented for the individual elements of an object, list, or message.

Examples of things considered values

u8

u16

u32

usize

i8

i16

i32

isize

bool

char

BitGravity

WinGravity

Protocol

KeySym

Examples of things not considered values

&str

String

&[u8]

Rect

Host

Required Methods

fn write_1b(self) -> Result<u8, WriteError>

Writes Self to a single byte (u8).

fn write_2b(self) -> Result<u16, WriteError>

Writes Self to two bytes (u16) using the system’s native endianness.

fn write_4b(self) -> Result<u32, WriteError>

Writes Self to four bytes (u32) using the system’s native endianness.

Provided Methods

fn write_1b_to(self, buf: &mut impl BufMut) -> Result<(), WriteError> where
Self: Sized,

fn write_2b_to(self, buf: &mut impl BufMut) -> Result<(), WriteError> where
Self: Sized,

fn write_4b_to(self, buf: &mut impl BufMut) -> Result<(), WriteError> where
Self: Sized,